generate-pdf-worker/requirements.txt
pika==1.3.2
redis==5.0.1
python-dotenv==1.0.1
boto3==1.34.0
requests==2.32.4
loguru
prometheus_client
Flask

